+++ CHANGELOG.md
@@ -23,6 +23,8 @@ All notable, unreleased changes to this project will be documented in this file.
 - Default product variant - #6140 by @tomaszszymanski129
 - Allow product variants to be sortable - #6138 by @tomaszszymanski129
 - Staff with only manage_orders should be able to query all stock related date. - #6139 by @fowczarek
+- Add filtering to `ProductVariants` query and option to fetch variant by sku in `ProductVariant` query - #6190 by @fowczarek
+
 
 ### Breaking Changes